Deprecating panos modules; pointing to Galaxy role (#52355)
* Deprecating panos modules; pointing to Galaxy role

* Updating porting guide; deprecating panos_set

* Changing 2.11 to 2.12
