+++ lib/ansible/modules/network/panos/_panos_match_rule.py
@@ -21,7 +21,7 @@
 #  limitations under the License.
 
 ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
+                    'status': ['deprecated'],
                     'supported_by': 'community'}
 
 DOCUMENTATION = '''
@@ -35,6 +35,10 @@ version_added: "2.5"
 requirements:
     - pan-python can be obtained from PyPI U(https://pypi.org/project/pan-python/)
     - pandevice can be obtained from PyPI U(https://pypi.org/project/pandevice/)
+deprecated:
+    alternative: Use U(https://galaxy.ansible.com/PaloAltoNetworks/paloaltonetworks) instead.
+    removed_in: "2.12"
+    why: Consolidating code base.
 notes:
     - Checkmode is not supported.
     - Panorama NOT is supported.
@@ -57,7 +61,10 @@ options:
     rule_type:
         description:
             - Type of rule. Valid types are I(security) or I(nat).
-        default: "security"
+        required: true
+        choices:
+            - security
+            - nat
     source_zone:
         description:
             - The source zone.
